Move proxy code to tools/claude_hooks/proxy/ subdirectory
Organize proxy-related code into a dedicated subdirectory for better
structure:

- Move auth_forwarding_proxy.py to proxy/auth_forwarding_proxy.py
- Move run_auth_proxy.py to proxy/run_auth_proxy.py
- Update BUILD.bazel to reflect new file locations
- Update proxy_setup.py to:
  - Use new module path (tools.claude_hooks.proxy.run_auth_proxy)
  - Use sys.executable instead of hardcoded "python"
- Update run_auth_proxy.py import to use new package path

This improves code organization by separating proxy implementation
from the main hooks logic. No __init__.py needed - Bazel auto-generates
stub __init__.py files via imports=[].
