+++ tools/claude_hooks/proxy/auth_forwarding_proxy.py
